Identify unimported locations using common code
